Add command-line argument handling for maximum moves
This commit adds command-line argument handling to the solver program. Now, the solver can be run with the following format:

./solver filename -m maxMoves

Users can now provide a specific maximum number of moves when running the solver. If the '-m' flag is not provided, the solver will use a default maximum number of moves (100 000).